Resolve "Make OUTPUT_BUFFER_SIZE configurable"

Closes #2392

See merge request meltano/meltano!1993

+++ src/meltano/core/runner/singer.py
@@ -15,7 +15,8 @@ from meltano.core.logging import capture_subprocess_output
 from meltano.core.plugin import PluginType
 from meltano.core.plugin.singer import PluginType, SingerTap, SingerTarget
 from meltano.core.plugin_invoker import PluginInvoker, invoker_factory
-from meltano.core.utils import file_has_data
+from meltano.core.project_settings_service import ProjectSettingsService
+from meltano.core.utils import human_size
 
 from . import Runner, RunnerError
 
@@ -48,13 +49,15 @@ class BookmarkWriter:
                 "Received state is invalid, incremental state has not been updated"
             )
 
-    def flush(self):
-        pass
-
 
 class SingerRunner(Runner):
     def __init__(self, elt_context: ELTContext):
         self.context = elt_context
+
+        self.project_settings_service = ProjectSettingsService(
+            self.context.project,
+            config_service=self.context.plugins_service.config_service,
+        )
 
     def stop(self, process, **wait_args):
         while True:
@@ -79,9 +82,17 @@ class SingerRunner(Runner):
         extractor_log = extractor_log or sys.stderr
         loader_log = loader_log or sys.stderr
 
+        # The StreamReader line length limit also acts as half the buffer size,
+        # which cannot be set directly:
+        # - https://github.com/python/cpython/blob/v3.8.7/Lib/asyncio/streams.py#L395-L396
+        # - https://github.com/python/cpython/blob/v3.8.7/Lib/asyncio/streams.py#L482
+        stream_buffer_size = self.project_settings_service.get("elt.buffer_size")
+        line_length_limit = stream_buffer_size // 2
+
         # Start tap
         try:
             p_tap = await tap.invoke_async(
+                limit=line_length_limit,
                 stdout=asyncio.subprocess.PIPE,  # Singer messages
                 stderr=asyncio.subprocess.PIPE,  # Log
             )
@@ -91,6 +102,7 @@ class SingerRunner(Runner):
         # Start target
         try:
             p_target = await target.invoke_async(
+                limit=line_length_limit,
                 stdin=asyncio.subprocess.PIPE,  # Singer messages
                 stdout=asyncio.subprocess.PIPE,  # Singer state
                 stderr=asyncio.subprocess.PIPE,  # Log
@@ -122,25 +134,71 @@ class SingerRunner(Runner):
             capture_subprocess_output(p_target.stderr, loader_log)
         )
 
-        # Wait for tap or target to complete
+        # Wait for tap or target to complete, or for one of the output handlers to raise an exception.
         tap_process_future = asyncio.ensure_future(p_tap.wait())
         target_process_future = asyncio.ensure_future(p_target.wait())
+        output_exception_future = asyncio.ensure_future(
+            asyncio.wait(
+                [
+                    tap_stdout_future,
+                    tap_stderr_future,
+                    target_stdout_future,
+                    target_stderr_future,
+                ],
+                return_when=asyncio.FIRST_EXCEPTION,
+            ),
+        )
+
         done, _ = await asyncio.wait(
-            [tap_process_future, target_process_future],
+            [tap_process_future, target_process_future, output_exception_future],
             return_when=asyncio.FIRST_COMPLETED,
         )
 
+        # If `output_exception_future` completes first, one of the output handlers raised an exception or all completed successfully.
+        if output_exception_future in done:
+            output_futures_done, _ = output_exception_future.result()
+            output_futures_failed = [
+                future
+                for future in output_futures_done
+                if future.exception() is not None
+            ]
+
+            if output_futures_failed:
+                # If any output handler raised an exception, re-raise it.
+
+                # Special behavior for the tap stdout handler raising a line length limit error.
+                if tap_stdout_future in output_futures_failed:
+                    self._handle_tap_line_length_limit_error(
+                        tap_stdout_future.exception(),
+                        line_length_limit=line_length_limit,
+                        stream_buffer_size=stream_buffer_size,
+                    )
+
+                failed_future = output_futures_failed.pop()
+                raise failed_future.exception()
+            else:
+                # If all of the output handlers completed without raising an exception,
+                # we still need to wait for the tap or target to complete.
+                done, _ = await asyncio.wait(
+                    [tap_process_future, target_process_future],
+                    return_when=asyncio.FIRST_COMPLETED,
+                )
+
         if target_process_future in done:
-            # If the target completes before the tap, it failed before processing all tap output
             target_code = target_process_future.result()
 
-            # Kill tap and cancel output processing since there's no more target to forward messages to
-            p_tap.kill()
-            tap_stdout_future.cancel()
-            tap_stderr_future.cancel()
+            if tap_process_future in done:
+                tap_code = tap_process_future.result()
+            else:
+                # If the target completes before the tap, it failed before processing all tap output
 
-            # Pretend the tap finished successfully since it didn't itself fail
-            tap_code = 0
+                # Kill tap and cancel output processing since there's no more target to forward messages to
+                p_tap.kill()
+                tap_stdout_future.cancel()
+                tap_stderr_future.cancel()
+
+                # Pretend the tap finished successfully since it didn't itself fail
+                tap_code = 0
 
             # Wait for all buffered target output to be processed
             await asyncio.wait([target_stdout_future, target_stderr_future])
@@ -200,3 +258,29 @@ class SingerRunner(Runner):
                 extractor_out=extractor_out,
                 loader_out=loader_out,
             )
+
+    def _handle_tap_line_length_limit_error(
+        self,
+        exception,
+        line_length_limit,
+        stream_buffer_size,
+    ):
+        # StreamReader.readline can raise a ValueError wrapping a LimitOverrunError:
+        # https://github.com/python/cpython/blob/v3.8.7/Lib/asyncio/streams.py#L549
+        if not isinstance(exception, ValueError):
+            return
+
+        exception = exception.__context__  # noqa: WPS609
+        if not isinstance(exception, asyncio.LimitOverrunError):
+            return
+
+        logging.error(
+            f"The extractor generated a message exceeding the message size limit of {human_size(line_length_limit)} (half the buffer size of {human_size(stream_buffer_size)})."
+        )
+        logging.error(
+            "To let this message be processed, increase the 'elt.buffer_size' setting to at least double the size of the largest expected message, and try again."
+        )
+        logging.error(
+            "To learn more, visit https://www.meltano.com/docs/settings.html#elt-buffer-size"
+        )
+        raise RunnerError("Output line length limit exceeded") from exception
